done deal

+++ minishell.h
@@ -6,7 +6,7 @@
 /*   By: hznagui <hznagui@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/01 12:17:32 by hznagui           #+#    #+#             */
-/*   Updated: 2023/04/05 00:55:34 by hznagui          ###   ########.fr       */
+/*   Updated: 2023/04/05 16:34:27 by hznagui          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ typedef struct data
     char *input;
     char **tab;
     int z;
-    int length;
+    size_t length;
     int lock;
     int lock1;
     size_t i;
